refactor(db): split workflow module by concern
