+++ packages/db/src/merge-tail.ts
@@ -218,7 +218,8 @@ const DEFENSE_EXACT = new Set([
   "scripts/merge-gate.sh",
   "scripts/merge-lease.sh",
   "scripts/regression-verification.sh",
-  "packages/db/src/workflow.ts",
+  "packages/db/src/chain-activation.ts",
+  "packages/db/src/inbox-decision.ts",
   "packages/db/src/merge-integrator.ts",
   "packages/db/src/gate-attestation.ts",
   "packages/db/src/merge-integrator-db.ts",
